Refactor Meta Ads presentation helpers
